+++ CHANGELOG.rst
@@ -2,8 +2,12 @@
 CHANGELOG
 =========
 
-1.12.1dev
-=========
+========
+1.13.dev
+========
+
+* feature: Estimator: add input mode to training channels
+* feature: Estimator: add model_uri and model_channel_name parameters
 * enhancement: Local Mode: support output_path. Can be either file:// or s3://
 
 1.12.0